api/pyproject.toml
[tool.poetry]
name = "movie-site"
version = "0.1.0"
description = ""
authors = ["Anton Kueltz <kueltz.anton@gmail.com>"]
readme = "README.md"
packages = [{include = "movie_site"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
strawberry-graphql = "^0.171.1"
sqlalchemy = "^2.0.9"
alembic = "^1.10.3"
argon2-cffi = "^21.3.0"
fastecdsa = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
ipython = "^8.12.0"
uvicorn = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
